feat: Change telegram_user_id column type from Integer to BigInteger in user_search_logs model

+++ backend/app/models/user_search_logs.py
@@ -2,7 +2,7 @@
 User search logs model for tracking mini-app search usage
 """
 
-from sqlalchemy import Column, Integer, String, DateTime, func, ForeignKey
+from sqlalchemy import Column, Integer, BigInteger, String, DateTime, func, ForeignKey
 from sqlalchemy.orm import relationship
 from app.core.database import Base
 
@@ -14,7 +14,7 @@ class UserSearchLog(Base):
 
     id = Column(Integer, primary_key=True, index=True)
     user_id = Column(Integer, ForeignKey("users.id"), nullable=False, index=True)
-    telegram_user_id = Column(Integer, nullable=False, index=True)
+    telegram_user_id = Column(BigInteger, nullable=False, index=True)
     search_query = Column(String(255), nullable=False)
     results_count = Column(Integer, default=0)
     searched_at = Column(DateTime(timezone=True), server_default=func.now(), nullable=False, index=True)
